Aşama15: Kullanıcı Yetkilendirme

+++ post/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect
+from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect, Http404
 from .models import Post
 from .forms import PostForm
 from django.contrib import messages
@@ -19,7 +19,8 @@ def post_detail(request, id):
 
 def post_create(request):
 
-    
+    if not request.user.is_authenticated():
+        return Http404()
 
     #if request.method == "POST":
     #    print(request.POST)
@@ -53,6 +54,9 @@ def post_create(request):
 
 
 def post_update(request, id):
+
+    if not request.user.is_authenticated():
+        return Http404()
     
     post = get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, request.FILES or None, instance=post)
@@ -70,6 +74,10 @@ def post_update(request, id):
 
 
 def post_delete(request, id):
-        post = get_object_or_404(Post, id=id)
-        post.delete()
-        return redirect('post:index')
+
+    if not request.user.is_authenticated():
+        return Http404()
+
+    post = get_object_or_404(Post, id=id)
+    post.delete()
+    return redirect('post:index')
